+++ ansible/roles/standardise/files/enamine_ro5/f40_create_stand_database.sql
@@ -3,19 +3,20 @@
  * Purpose: Creates fragmentation company specific tables
  *
  * Author | Date    | Version
+ * Alan   | 10/2024 | BIGINT Version
  * Duncan | 03/2020 | Initial Version
  *
  */
 
 
 /*
- * Drop all tables if they exist (NB Order because of table key constrants) 
+ * Drop all tables if they exist (NB Order because of table key constrants)
  */
 
 DROP TABLE IF EXISTS i_mols_enamine;
 
 /*
- * Create i_mols        
+ * Create i_mols
  */
 CREATE TABLE i_mols_enamine (
   osmiles TEXT,
@@ -24,6 +25,6 @@ CREATE TABLE i_mols_enamine (
   hac SMALLINT,
   rac SMALLINT,
   cmpd_id TEXT,
-  isomol_id INTEGER,
-  nonisomol_id INTEGER
+  isomol_id bigint,
+  nonisomol_id bigint
 );
